#42 feat: Create an interface that enforces predictable operator names
Create an interface that enforces predictable operator names for the Decodable stream sources and sinks. Currently, the Decodable stream sources and sinks, when inserted into a DataStream graph, are arbitrarily nameable. This presents issues for the Decodable platform when attempting to correlate runtime metrics from the graph operators back to the referenced streams.